requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
discord.py>=2.0.0
httpx>=0.23.0
jinja2>=3.0.0
python-dotenv>=0.19.0
itsdangerous>=2.0.0